requirements.txt
python-dotenv
rich
textual
typing_extensions
